+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `formService.parseFreshdeskFieldFormElementOptions()`
 - `formService.getFormElementOptionsSetOptions()`
 - `formService.loadFormElementDynamicOptions()`
+- `runLookupOnClear` and `records` to response from `formService.getFormElementLookups()`
 
 ### Changed